+++ packages/analytics-js-service-worker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.2.25](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.24...@rudderstack/analytics-js-service-worker@3.2.25) (2025-05-09)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.19.0`
 ## [3.2.24](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-service-worker@3.2.23...@rudderstack/analytics-js-service-worker@3.2.24) (2025-04-25)
 
 ### Dependency Updates